feat: separate rule & rules mod - so that folder_iso_struct does not have to handle generating pub mod, which failed -
