+++ cli/src/lib.rs
@@ -2,4 +2,5 @@ pub mod git;
 pub mod message;
 pub mod result;
 pub mod rule;
+pub mod rules;
 pub mod settings;
