+++ cli/src/rule/body_empty.rs
@@ -1,5 +1,5 @@
-use super::Level;
-use crate::{make_rule, message::Message, result::Violation, rule::Rule};
+use crate::rules::Level;
+use crate::{make_rule, message::Message, result::Violation, rules::Rule};
 
 make_rule! {
     BodyEmpty,
