+++ cli/src/rule/subject_empty.rs
@@ -1,6 +1,6 @@
-use crate::{make_rule, message::Message, result::Violation, rule::Rule};
+use crate::{make_rule, message::Message, result::Violation, rules::Rule};
 
-use super::Level;
+use crate::rules::Level;
 
 make_rule! {
     SubjectEmpty,
