+++ cli/src/rule/type.rs
@@ -1,6 +1,6 @@
-use crate::{make_options_rule, message::Message, result::Violation, rule::Rule};
+use crate::{make_options_rule, message::Message, result::Violation, rules::Rule};
 
-use super::Level;
+use crate::rules::Level;
 make_options_rule! {
     Type,
     "type",
