+++ cli/src/rules.rs
@@ -1,22 +1,6 @@
-use crate::{message::Message, result::Violation};
+use crate::{message::Message, result::Violation, rule::*};
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
-pub mod body_empty;
-pub mod body_max_length;
-pub mod description_empty;
-pub mod description_format;
-pub mod description_max_length;
-pub mod footers_empty;
-pub mod scope;
-pub mod scope_empty;
-pub mod scope_format;
-pub mod scope_max_length;
-pub mod subject_empty;
-pub mod r#type;
-pub mod type_empty;
-pub mod type_format;
-pub mod type_max_length;
-
 include!(concat!(env!("OUT_DIR"), "/Rules.rs"));
 
 /// Rule is a collection of rules.
